+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Sastre 1.16 [October 15, 2021]
 - Attach/detach tasks now on both Sastre and Sastre-Pro.
 - Infra changes related to the report task, not user-facing.
 - Report task diff option now allow specific sections to not be considered for diff comparison. Via skip_diff option in the report specification.
+- [#105] Tasks with --dryrun option, when executed without --verbose, now display a dryrun action preview at the end.
 
 
 Sastre 1.15 [September 27, 2021]